BZ2983-Tables for supported enumeration per device
This is an editorial CR of OCF Device Spec v2.0.3.
This CR proposes to add tables that list the supported enumeration and description depending to devices. (The supported enumerations include the new values of BZ#2860 and BZ#2915.) Particularly, the device types included in the proposed CR are only 8 devices types following as; Air Conditioner, Air Purifier, Dishwasher, Dryer, Oven, Robot Cleaner, Steam Closet, and Washer.